rename doveauth-dictproxy to doveauth
